scripts/tdr/twisttcap/requirements.txt
firecloud==0.16.31
google-cloud-storage==1.33.0
google-cloud-bigquery==2.30.1
oauth2client==4.1.3
pandas==1.1.4
pip==19.3.1
tenacity==5.1.1